Libft V1.5

+++ Libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: eguelin <eguelin@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/07 17:20:15 by eguelin           #+#    #+#             */
-/*   Updated: 2022/11/10 04:21:02 by eguelin          ###   ########lyon.fr   */
+/*   Updated: 2022/11/17 18:38:20 by eguelin          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	pos;
+	size_t	pos;
 
 	pos = 0;
 	while (s[pos])
